+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/exception/GlobalEventException.java
@@ -1,0 +1,16 @@
+package io.github.ulisse1996.jaorm.exception;
+
+public class GlobalEventException extends RuntimeException {
+
+    GlobalEventException(String message) {
+        super(message);
+    }
+
+    GlobalEventException(Throwable throwable) {
+        super(throwable);
+    }
+
+    GlobalEventException(String message, Throwable ex) {
+        super(message, ex);
+    }
+}