+++ jaorm-core/src/test/java/io/github/ulisse1996/jaorm/BaseDaoTest.java
@@ -316,6 +316,8 @@ class BaseDaoTest {
                     .thenReturn(true);
             mkEntity.when(() -> EntityEvent.forType(eventType))
                     .thenReturn(event);
+            Mockito.when(event.applyAndReturn(Mockito.any()))
+                    .then(i -> i.getArgument(0));
             switch (eventType) {
                 case PERSIST:
                     dao.insert(new DelegatesMock.MyEntity());