+++ jaorm-core/src/test/java/io/github/ulisse1996/jaorm/QueryRunnerTest.java
@@ -37,6 +37,13 @@ class QueryRunnerTest {
     }
 
     @Test
+    void should_return_same_rows() {
+        MockedRunner runner = new MockedRunner();
+        runner.registerUpdatedRows(1, 10);
+        Assertions.assertEquals(10, runner.getUpdatedRows(1));
+    }
+
+    @Test
     void should_not_find_simple_query_runner() {
         try (MockedStatic<ServiceFinder> mk = Mockito.mockStatic(ServiceFinder.class)) {
             mk.when(() -> ServiceFinder.loadServices(QueryRunner.class))